django_backend/requirements.txt
# Django==5.2.9
aiogram>=3.0.0
python-dotenv==1.1.1
qdrant-client
mysqlclient
djangorestframework
google-generativeai
djangorestframework-simplejwt
django-redis
redis
channels
daphne
channels-redis